+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+
+* XX.X.X
+  - Added wavelet operator, wrapping PyWavelets operator as a CIL operator
+  - Added L1Sparsity function, allowing calculations of `|Ax-b|_1` and it's proximal, in the case of orthogonal operators, A
+  - Bug fix in gradient descent `update_objective` called twice on the initial point.
+  - ProjectionMap operator bug fix in adjoint and added documentation #1743
+
+
 * 24.0.0
   - Update to new CCPi-Regularisation toolkit v24.0.0. This is a backward incompatible release of the toolkit.
   - CIL plugin support for TIGRE version v2.6